+++ src/refactoring/components/features/product/ProductCard.tsx
@@ -1,63 +1,26 @@
 import { Product } from "../../../../types";
-import { useCartContext } from "../../../contexts/CartContext";
-import { getRemainingStock } from "../../../models/product";
-import { getMaxDiscount } from "../../../utils";
 
 type ProductCardProps = {
   product: Product;
+  onEdit: (product: Product) => void;
 };
 
-export const ProductCard = ({ product }: ProductCardProps) => {
-  const { cart, addToCart } = useCartContext();
-
-  const remainingStock = getRemainingStock(cart, product);
-
+export const ProductCard = ({ product, onEdit }: ProductCardProps) => {
   return (
-    <div
-      key={product.id}
-      data-testid={`product-${product.id}`}
-      className="bg-white p-3 rounded shadow"
-    >
-      <div className="flex justify-between items-center mb-2">
-        <span className="font-semibold">{product.name}</span>
-        <span className="text-gray-600">
-          {product.price.toLocaleString()}원
-        </span>
-      </div>
-      <div className="text-sm text-gray-500 mb-2">
-        <span
-          className={`font-medium ${
-            remainingStock > 0 ? "text-green-600" : "text-red-600"
-          }`}
-        >
-          재고: {remainingStock}개
-        </span>
-        {product.discounts.length > 0 && (
-          <span className="ml-2 font-medium text-blue-600">
-            최대 {(getMaxDiscount(product.discounts) * 100).toFixed(0)}% 할인
+    <div>
+      {product.discounts.map((discount, index) => (
+        <div key={index} className="mb-2">
+          <span>
+            {discount.quantity}개 이상 구매 시 {discount.rate * 100}% 할인
           </span>
-        )}
-      </div>
-      {product.discounts.length > 0 && (
-        <ul className="list-disc list-inside text-sm text-gray-500 mb-2">
-          {product.discounts.map((discount, index) => (
-            <li key={index}>
-              {discount.quantity}개 이상: {(discount.rate * 100).toFixed(0)}%
-              할인
-            </li>
-          ))}
-        </ul>
-      )}
+        </div>
+      ))}
       <button
-        onClick={() => addToCart(product)}
-        className={`w-full px-3 py-1 rounded ${
-          remainingStock > 0
-            ? "bg-blue-500 text-white hover:bg-blue-600"
-            : "bg-gray-300 text-gray-500 cursor-not-allowed"
-        }`}
-        disabled={remainingStock <= 0}
+        data-testid="modify-button"
+        onClick={() => onEdit(product)}
+        className="bg-blue-500 text-white px-2 py-1 rounded hover:bg-blue-600 mt-2"
       >
-        {remainingStock > 0 ? "장바구니에 추가" : "품절"}
+        수정
       </button>
     </div>
   );
