+++ src/refactoring/components/features/product/ProductEditForm.tsx
@@ -4,6 +4,8 @@ import { Accordion } from "../../ui/common/Accordion";
 import { formatCurrency } from "../../../utils";
 import { Discount, Product } from "../../../../types";
 import { FormFieldInput } from "../../ui/common/FormFieldInput";
+import { ProductCard } from "./ProductCard";
+import { DiscountEditor } from "./DiscountEditor";
 
 export const ProductEditForm = () => {
   const { products, updateProduct } = useProductContext();
@@ -127,58 +129,15 @@ export const ProductEditForm = () => {
                 onChange={handleInputChange}
               />
               {/* 할인 정보 수정 부분 */}
-              <div>
-                <h4 className="text-lg font-semibold mb-2">할인 정보</h4>
-                {editingProduct.discounts.map((discount, index) => (
-                  <div
-                    key={index}
-                    className="flex justify-between items-center mb-2"
-                  >
-                    <span>
-                      {discount.quantity}개 이상 구매 시 {discount.rate * 100}%
-                      할인
-                    </span>
-                    <button
-                      onClick={() => handleRemoveDiscount(product.id, index)}
-                      className="bg-red-500 text-white px-2 py-1 rounded hover:bg-red-600"
-                    >
-                      삭제
-                    </button>
-                  </div>
-                ))}
-                <div className="flex space-x-2">
-                  <input
-                    type="number"
-                    placeholder="수량"
-                    value={newDiscount.quantity}
-                    onChange={(e) =>
-                      setNewDiscount({
-                        ...newDiscount,
-                        quantity: parseInt(e.target.value),
-                      })
-                    }
-                    className="w-1/3 p-2 border rounded"
-                  />
-                  <input
-                    type="number"
-                    placeholder="할인율 (%)"
-                    value={newDiscount.rate * 100}
-                    onChange={(e) =>
-                      setNewDiscount({
-                        ...newDiscount,
-                        rate: parseInt(e.target.value) / 100,
-                      })
-                    }
-                    className="w-1/3 p-2 border rounded"
-                  />
-                  <button
-                    onClick={() => handleAddDiscount(product.id)}
-                    className="w-1/3 bg-blue-500 text-white p-2 rounded hover:bg-blue-600"
-                  >
-                    할인 추가
-                  </button>
-                </div>
-              </div>
+              <DiscountEditor
+                discounts={editingProduct.discounts}
+                newDiscount={newDiscount}
+                onNewDiscountChange={setNewDiscount}
+                onAddDiscount={() => handleAddDiscount(product.id)}
+                onRemoveDiscount={(index) =>
+                  handleRemoveDiscount(product.id, index)
+                }
+              />
               <button
                 onClick={handleEditComplete}
                 className="bg-green-500 text-white px-2 py-1 rounded hover:bg-green-600 mt-2"
@@ -187,23 +146,7 @@ export const ProductEditForm = () => {
               </button>
             </div>
           ) : (
-            <div>
-              {product.discounts.map((discount, index) => (
-                <div key={index} className="mb-2">
-                  <span>
-                    {discount.quantity}개 이상 구매 시 {discount.rate * 100}%
-                    할인
-                  </span>
-                </div>
-              ))}
-              <button
-                data-testid="modify-button"
-                onClick={() => handleEditProduct(product)}
-                className="bg-blue-500 text-white px-2 py-1 rounded hover:bg-blue-600 mt-2"
-              >
-                수정
-              </button>
-            </div>
+            <ProductCard product={product} onEdit={handleEditProduct} />
           )}
         </Accordion>
       ))}
