+++ src/refactoring/components/features/product/ProductList.tsx
@@ -1,5 +1,5 @@
 import { useProductContext } from "../../../contexts/ProductContext";
-import { ProductCard } from "./ProductCard";
+import { ProductItem } from "./ProductItem";
 
 export const ProductList = () => {
   const { products } = useProductContext();
@@ -9,7 +9,7 @@ export const ProductList = () => {
       <h2 className="text-2xl font-semibold mb-4">상품 목록</h2>
       <div className="space-y-2">
         {products.map((product, index) => (
-          <ProductCard product={product} key={index} />
+          <ProductItem product={product} key={index} />
         ))}
       </div>
     </div>
